Consolidated Dinosaurs in adoption page into one larger section
