+++ client/src/index.css
@@ -65,7 +65,7 @@ footer {
   margin-bottom: 20px;
 }
 
-#carnivores {
+#dinosaursAdopt {
   margin-top: 70px;
   position: relative;
   bottom: 30px;
